refactor: extract interceptor
